Conflicts:
	src/server/game/Spells/Spell.cpp

+++ src/server/game/AI/SmartScripts/SmartScriptMgr.h
@@ -885,7 +885,7 @@ struct SmartAction
 
         struct
         {
-            uint8 pointId;
+            uint32 pointId;
         } MoveToPos;
 
         struct
